requirements.txt
google-api-python-client==2.131.0
python-telegram-bot==21.2
python-dotenv==1.0.1